+++ src/nbsap/views/objectives.py
@@ -168,14 +168,15 @@ def edit_national_objective(request, pk=None, parent=None):
 
 @auth_required
 def delete_national_objective(request, pk):
-    objective = get_object_or_404(models.NationalObjective, pk=pk)
-    parent = objective.parent
-    objective.delete()
-    messages.success(request, _('Objective successfully deleted.') + "")
-    if parent:
-        return redirect('view_national_objective', pk=parent.pk)
-    else:
-        return redirect('list_national_objectives')
+    if request.method == 'POST':
+        objective = get_object_or_404(models.NationalObjective, pk=pk)
+        parent = objective.parent
+        objective.delete()
+        messages.success(request, _('Objective successfully deleted.') + "")
+        if parent:
+            return redirect('view_national_objective', pk=parent.pk)
+        else:
+            return redirect('list_national_objectives')
 
 
 def get_national_objective_title(request, pk=None):
